requirements.txt
# Core deep learning
torch>=2.0.0 --index-url https://download.pytorch.org/whl/nightly/cu129
torchvision>=0.15.0 --index-url https://download.pytorch.org/whl/nightly/cu129
torchaudio>=2.0.0 --index-url https://download.pytorch.org/whl/nightly/cu129

# YOLOv8
ultralytics>=8.0.0

# Data handling
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2

# Image & video
opencv-python>=4.7
Pillow>=9.4

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Utils
pyyaml>=6.0

# Optional
jupyter>=1.0.0
notebook>=6.5